Created own component for backup-table
