+++ apps/frontend/src/app/backups-overview/component/backup-table/backup-table.component.html
@@ -1,0 +1,168 @@
+<div>
+  <h3>All Backups:</h3>
+  <clr-datagrid
+    *ngIf="backups$ | async as backups"
+    (clrDgRefresh)="refresh($event)"
+    [clrDgLoading]="loading"
+  >
+    <clr-dg-column [clrDgField]="'saveset'">
+      <ng-container *clrDgHideableColumn>
+        Saveset
+        <clr-dg-filter [clrDgFilter]="backupSavesetFilter">
+          <clr-input-container>
+            <label>Saveset:</label>
+            <input
+              clrInput
+              type="string"
+              [(ngModel)]="backupSavesetFilter.ranges.saveset"
+              (ngModelChange)="
+                  backupSavesetFilter.updateRanges({ saveset: $event })
+                "
+            />
+          </clr-input-container>
+        </clr-dg-filter>
+      </ng-container>
+    </clr-dg-column>
+
+    <clr-dg-column [clrDgField]="'sizeMB'">
+      <ng-container *clrDgHideableColumn>
+        Size (MB)
+        <clr-dg-filter [clrDgFilter]="backupSizeFilter">
+          <clr-input-container>
+            <label>From:</label>
+            <input
+              clrInput
+              type="number"
+              [(ngModel)]="backupSizeFilter.ranges.fromSizeMB"
+              (ngModelChange)="
+                  backupSizeFilter.updateRanges({ fromSizeMB: $event })
+                "
+            />
+          </clr-input-container>
+          <clr-input-container>
+            <label>To:</label>
+            <input
+              clrInput
+              type="number"
+              [(ngModel)]="backupSizeFilter.ranges.toSizeMB"
+              (ngModelChange)="
+                  backupSizeFilter.updateRanges({ toSizeMB: $event })
+                "
+            />
+          </clr-input-container>
+        </clr-dg-filter>
+      </ng-container>
+    </clr-dg-column>
+
+    <clr-dg-column [clrDgField]="'taskName'">
+      <ng-container *clrDgHideableColumn>
+        Task
+        <clr-dg-filter [clrDgFilter]="taskFilter">
+          <clr-input-container>
+            <label>Task:</label>
+            <input
+              clrInput
+              type="string"
+              [(ngModel)]="taskFilter.ranges.taskName"
+              (ngModelChange)="taskFilter.updateRanges({ taskName: $event })"
+            />
+          </clr-input-container>
+        </clr-dg-filter>
+      </ng-container>
+    </clr-dg-column>
+
+    <clr-dg-column [clrDgField]="'type'">
+      <ng-container *clrDgHideableColumn>
+        Type
+        <clr-dg-filter [clrDgFilter]="typeFilter">
+          <clr-combobox-container>
+            <clr-combobox
+              [(ngModel)]="selectedBackupTypesForTable"
+              [clrMulti]="true"
+              (ngModelChange)="setBackupTypesForTable($event)"
+              placeholder="Select backup types">
+              <ng-container *clrOptionSelected="let value">
+                {{ value }}
+              </ng-container>
+              <clr-options>
+                <clr-option *ngFor="let option of backupEnumTypes" [clrValue]="option">
+                  {{ option }}
+                </clr-option>
+              </clr-options>
+            </clr-combobox>
+          </clr-combobox-container>
+        </clr-dg-filter>
+      </ng-container>
+    </clr-dg-column>
+
+    <clr-dg-column
+      [clrDgField]="'creationDate'"
+      [clrDgSortOrder]="ClrDatagridSortOrder.DESC"
+    >
+      <ng-container *clrDgHideableColumn>
+        Creation Date
+        <clr-dg-filter [clrDgFilter]="backupDateFilter">
+          <clr-input-container>
+            <label>From:</label>
+            <input
+              clrInput
+              type="date"
+              [(ngModel)]="backupDateFilter.ranges.fromDate"
+              (ngModelChange)="
+                  backupDateFilter.updateRanges({ fromDate: $event })
+                "
+            />
+          </clr-input-container>
+          <clr-input-container>
+            <label>To:</label>
+            <input
+              clrInput
+              type="date"
+              [(ngModel)]="backupDateFilter.ranges.toDate"
+              (ngModelChange)="
+                  backupDateFilter.updateRanges({ toDate: $event })
+                "
+            />
+          </clr-input-container>
+        </clr-dg-filter>
+      </ng-container>
+    </clr-dg-column>
+
+    <clr-dg-placeholder>We couldn't find any backups!</clr-dg-placeholder>
+    <clr-dg-row *ngFor="let backup of backups.data" [clrDgItem]="backup">
+      <clr-dg-cell>{{ backup.saveset }}</clr-dg-cell>
+      <clr-dg-cell>{{ backup.sizeMB }}</clr-dg-cell>
+      <clr-dg-cell>{{ backup.taskId?.displayName ?? '' }}</clr-dg-cell>
+      <clr-dg-cell>{{ backup.type }}</clr-dg-cell>
+      <clr-dg-cell>{{ backup.creationDate | date }}</clr-dg-cell>
+      <clr-dg-action-overflow>
+        <button class="action-item">More Details</button>
+        <button class="action-item">Edit</button>
+        <button class="action-item">Delete</button>
+      </clr-dg-action-overflow>
+    </clr-dg-row>
+
+    <clr-dg-footer>
+      <clr-dg-pagination
+        #pagination
+        [clrDgTotalItems]="backups.paginationData.total"
+      >
+        <clr-dg-page-size [clrPageSizeOptions]="[10, 20, 50, 100]">
+          Backups per page
+        </clr-dg-page-size>
+        {{
+          backups.paginationData.offset
+            ? +backups.paginationData.offset + 1
+            : 0
+        }}
+        -
+        {{
+          backups.paginationData.offset
+            ? +backups.paginationData.offset + pageSize
+            : 0
+        }}
+        of {{ backups.paginationData.total }} backups
+      </clr-dg-pagination>
+    </clr-dg-footer>
+  </clr-datagrid>
+</div>
